* XML I/O formatting and indentation rewritted. The former implementation broke a few tests. * Changed a few calls to WriteStr to WriteUTF8Str in order to avoid warnings and better support unicode identifiers; added TInstantWriter.WriteUTF8Str for D<2009.

+++ Source/Core/InstantConsts.pas
@@ -79,7 +79,7 @@ const
 {$ENDIF}
 
 var
-  InstantXMLIndentationSize: Integer = 2;
+  InstantXMLIndentationSize: Byte = 2;
 
 resourcestring
   SAccessError = 'Cannot access attribute %s(''%s'') as type: %s';
